// src/hunterxhunter/models/HunterXHunterTags.ts
export interface HunterXHunterTags {
  abilities: string;
  abilities_description: string;
  affiliation: string;
  age: string;
  alias: string;
  birthday: string;
  blood_type: string;
  description: string;
  eye_color: string;
  gender: string;
  hair_color: string;
  japanese_voice: string;
  height: string;
  name: string;
  nen_type: string;
  occupation: string;
  relatives: string;
  photo: string;
  status: string;
  weight: string;
}